requirements.txt
Django==4.2.16
djangorestframework==3.15.2
drf_yasg==1.21.8
setuptools==58.0.4
setuptools==75.1.0
